requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
kucoin-python==1.0.9
pip==21.1.2
requests==2.26.0
setuptools==57.0.0
urllib3==1.26.7
websockets==10.1
wheel==0.36.2